Moved Trampoline to Dispatchers per Paco's suggestion
