+++ modules/effects/arrow-effects-data/src/main/kotlin/arrow/effects/typeclasses/Dispatchers.kt
@@ -4,4 +4,5 @@ import kotlin.coroutines.CoroutineContext
 
 interface Dispatchers<F> {
   fun default(): CoroutineContext
+  fun trampoline(): CoroutineContext
 }
