+++ modules/effects/arrow-effects-data/src/main/kotlin/arrow/effects/typeclasses/suspended/FxSyntax.kt
@@ -12,8 +12,8 @@ import arrow.effects.typeclasses.*
 import arrow.typeclasses.ApplicativeError
 import arrow.typeclasses.Monad
 import arrow.typeclasses.suspended.BindSyntax
-import java.util.concurrent.ForkJoinPool
-import kotlin.coroutines.*
+import kotlin.coroutines.CoroutineContext
+import kotlin.coroutines.startCoroutine
 
 interface FxSyntax<F> : Concurrent<F>, BindSyntax<F> {
 
@@ -21,6 +21,9 @@ interface FxSyntax<F> : Concurrent<F>, BindSyntax<F> {
 
   val NonBlocking: CoroutineContext
     get() = dispatchers().default()
+
+  val Trampoline: CoroutineContext
+    get() = dispatchers().trampoline()
 
   fun <A, B> CoroutineContext.parTraverse(
     effects: Iterable<Kind<F, A>>,
@@ -128,38 +131,4 @@ interface FxSyntax<F> : Concurrent<F>, BindSyntax<F> {
   fun <A> Iterable<Iterable<suspend () -> A>>.flatSequence(): Kind<F, List<A>> =
     flatten().sequence()
 
-}
-
-val Trampoline: CoroutineContext = EmptyCoroutineContext + TrampolinePool(ForkJoinPool())
-
-val iterations: ThreadLocal<Int> = ThreadLocal.withInitial { 0 }
-
-private class TrampolinePool(
-  val pool: ForkJoinPool,
-  val asyncBoundaryAfter: Int = 127
-) : AbstractCoroutineContextElement(ContinuationInterceptor), ContinuationInterceptor {
-
-  override fun <T> interceptContinuation(continuation: Continuation<T>): Continuation<T> =
-    TrampolinedContinuation(pool, continuation.context.fold(continuation) { cont, element ->
-      if (element != this@TrampolinePool && element is ContinuationInterceptor)
-        element.interceptContinuation(cont) else cont
-    }, asyncBoundaryAfter)
-}
-
-private class TrampolinedContinuation<T>(
-  val pool: ForkJoinPool,
-  val cont: Continuation<T>,
-  val asyncBoundaryAfter: Int
-) : Continuation<T> {
-  override val context: CoroutineContext = cont.context
-
-  override fun resumeWith(result: Result<T>) {
-    if (iterations.get() > asyncBoundaryAfter) {
-      iterations.set(0)
-      pool.execute { cont.resumeWith(result) }
-    } else {
-      iterations.set(iterations.get() + 1)
-      cont.resumeWith(result)
-    }
-  }
 }
