+++ modules/effects/arrow-effects-io-extensions/src/main/kotlin/arrow/effects/IODispatchers.kt
@@ -28,4 +28,38 @@ object IODispatchers {
       pool.execute { cont.resumeWith(result) }
     }
   }
+
+  val TrampolinePool: CoroutineContext = EmptyCoroutineContext + TrampolinePoolElement(ForkJoinPool())
+
+  val iterations: ThreadLocal<Int> = ThreadLocal.withInitial { 0 }
+
+  private class TrampolinePoolElement(
+    val pool: ForkJoinPool,
+    val asyncBoundaryAfter: Int = 127
+  ) : AbstractCoroutineContextElement(ContinuationInterceptor), ContinuationInterceptor {
+
+    override fun <T> interceptContinuation(continuation: Continuation<T>): Continuation<T> =
+      TrampolinedContinuation(pool, continuation.context.fold(continuation) { cont, element ->
+        if (element != this@TrampolinePoolElement && element is ContinuationInterceptor)
+          element.interceptContinuation(cont) else cont
+      }, asyncBoundaryAfter)
+  }
+
+  private class TrampolinedContinuation<T>(
+    val pool: ForkJoinPool,
+    val cont: Continuation<T>,
+    val asyncBoundaryAfter: Int
+  ) : Continuation<T> {
+    override val context: CoroutineContext = cont.context
+
+    override fun resumeWith(result: Result<T>) {
+      if (iterations.get() > asyncBoundaryAfter) {
+        iterations.set(0)
+        pool.execute { cont.resumeWith(result) }
+      } else {
+        iterations.set(iterations.get() + 1)
+        cont.resumeWith(result)
+      }
+    }
+  }
 }
